refactor: split large main.py into modular components and update imports
